refactor: MovieService 분리, WebClient 사용
비동기적 처리에 대해서는 추후 고민
