+++ src/main/java/com/personal/movie/controller/MovieController.java
@@ -2,10 +2,7 @@ package com.personal.movie.controller;
 
 import com.personal.movie.dto.MovieDto;
 import com.personal.movie.service.MovieService;
-import java.io.IOException;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,7 +11,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -23,9 +19,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class MovieController {
 
     private final MovieService movieService;
-
-    @Value("${movie.secret}")
-    private String apiKey;
 
     @PostMapping
     public ResponseEntity<MovieDto> insertMovie(@RequestBody MovieDto movieDto) {
@@ -46,24 +39,6 @@ public class MovieController {
     @DeleteMapping("/{id}")
     public ResponseEntity<MovieDto> deleteMovie(@PathVariable Long id) {
         return ResponseEntity.ok(movieService.deleteMovie(id));
-    }
-
-    @PostMapping("/popular")
-    public ResponseEntity<List<MovieDto>> getPopularMovies() {
-        return ResponseEntity.ok(movieService.getPopularMovies(apiKey));
-    }
-
-    @PostMapping("/keyword")
-    public ResponseEntity<List<MovieDto>> getMovieByKeyword(@RequestParam String keyword)
-        throws IOException {
-        return ResponseEntity.ok(movieService.getMovieByKeyword(apiKey, keyword));
-    }
-
-    @PostMapping("/person")
-    public ResponseEntity<List<MovieDto>> getMovieByPerson(@RequestParam String personName)
-        throws IOException {
-
-        return ResponseEntity.ok(movieService.getMovieByPerson(apiKey, personName));
     }
 
 }
